webmidi/requestmidiaccess.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="UTF-8">
    <title>Web MIDI API Test: requestMIDIAccess()</title>
    <link rel="author" title="Ryoya KAWAI" href="mailto:ryoya.kawai@gmail.com">
    <link rel="help" href="http://webaudio.github.io/web-midi-api/#requestmidiaccess"> <!-- 4.1 -->
    <link rel="help" href="http://webaudio.github.io/web-midi-api/#midiaccess-interface"> <!-- 5 -->
    <meta name="assert" content="Return true when test success in requestMIDIAccess().">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <p>Success when requestMIDIAccess() exists and MIDIAccess Interface is correctly obtained .</p>
    <div id="log"></div>
    <!-- PAGE CONTENT -->
    <script type="text/javascript">
    var MIDI;
    var test_requestMIDIAccess = async_test("requestMIDIAccess({sysex: flase})");
    // [TODO] need more test on the MIDIOptions flag
    function testWMA() {
        var ret=navigator.requestMIDIAccess({sysex:false}).then( successCallback, errorCallback );
        // 4.1 requestMIDIAccess() (in { sysex:false })
        test(function(){
            assert_class_string(ret, "Promise");
        }, "4.1 Return type in {sysex:false}");
        function successCallback(access) {
            MIDI=access;
            // 5. MIDIAccess Interface
            test_requestMIDIAccess.step(function() {
                test(function(){
                    assert_class_string(MIDI, "MIDIAccess");
                }, "5 MIDIAccess Interface name test in {sysex:false}");
                // 5.1 Attributes
                test(function(){
                    assert_class_string(MIDI.inputs, "MIDIInputMap");
                    assert_class_string(MIDI.outputs, "MIDIOutputMap");
                    //assert_true(typeof MIDI.onstatechange=="object");
                    assert_equals("boolean", typeof MIDI.sysexEnabled);
                    assert_false(MIDI.sysexEnabled);
                }, "5.1 MIDIAccess interface attribute type test in {sysex:false}.");
                test(function(){
                    assert_idl_attribute(MIDI, "inputs");
                    assert_idl_attribute(MIDI, "outputs");
                    assert_idl_attribute(MIDI, "onstatechange");
                    assert_idl_attribute(MIDI, "sysexEnabled");
                }, "5.1 MIDIAccess interface attribute name test in {sysex:false}.");
                test_requestMIDIAccess.done();
            });
        }
        function errorCallback(message) {
            test_requestMIDIAccess.step(function() {
                assert_equals("SecurityError", message.name);
                test_requestMIDIAccess.done();
            });
        }
    }
    testWMA();

    MIDI="";
    var test_requestMIDIAccess02 = async_test("requestMIDIAccess({sysex: true}) - W3C Spec: Web MIDI API");
    function testWMA02() {
        var ret02=navigator.requestMIDIAccess({sysex:true}).then( successCallback, errorCallback );
        // 4.1 requestMIDIAccess() (in { sysex:true })
        test(function(){
            assert_class_string(ret02, "Promise");
        }, "Return type test in sysex:true");
        function successCallback(access) {
            MIDI=access;
            // 5. MIDIAccess Interface
            test_requestMIDIAccess02.step(function() {
                test(function(){
                    assert_class_string(MIDI, "MIDIAccess");
                }, "MIDIAccess interface name test in {sysex:true}");
                // 5.1 Attributes
                test(function(){
                    assert_class_string(MIDI.inputs, "MIDIInputMap");
                    assert_class_string(MIDI.outputs, "MIDIOutputMap");
                    assert_true(typeof MIDI.onstatechange=="object");
                    assert_equals("boolean", typeof MIDI.sysexEnabled);
                    assert_true(MIDI.sysexEnabled);
                }, "5.1 MIDIAccess interface attribute type test in {sysex:true}.");
                test(function(){
                    assert_idl_attribute(MIDI, "inputs");
                    assert_idl_attribute(MIDI, "outputs");
                    assert_idl_attribute(MIDI, "onstatechange");
                    assert_idl_attribute(MIDI, "sysexEnabled");
                }, "5.1 MIDIAccess interface attribute name test in {sysex:true}.");
            });
            test_requestMIDIAccess02.done();
        }
        function errorCallback(message) {
            test_requestMIDIAccess02.step(function() {
                assert_equals("SecurityError", message.name, "");
                test_requestMIDIAccess02.done();
            });
        }
    }
    testWMA02();
    </script>
  </body>
</html>
